backend/requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
passlib[bcrypt]
fastapi[all]     
pyjwt
opencv-python
numpy
easyocr
ultralytics
pyserial